lets use floats, why not

+++ src/slate.js
@@ -63,8 +63,8 @@ var Resolvers = {
   greaterThan:{
     requiresOperand:true,
     fn:function(target, test){
-      target = parseInt(target, 10);
-      test = parseInt(test, 10);
+      target = parseFloat(target, 10);
+      test = parseFloat(test, 10);
       expect(target, 'Assertion target').to.be.ok;
       expect(target, 'Assertion target').to.be.a('number');
       expect(test, 'Operand').to.be.a('number');
@@ -74,8 +74,8 @@ var Resolvers = {
   lessThan:{
     requiresOperand:true,
     fn:function(target, test){
-      target = parseInt(target, 10);
-      test = parseInt(test, 10);
+      target = parseFloat(target, 10);
+      test = parseFloat(test, 10);
       expect(target, 'Assertion target').to.be.ok;
       expect(target, 'Assertion target').to.be.a('number');
       expect(test, 'Operand').to.be.a('number');
